chore: Switch to a class and rename for clarity
