+++ packages/runed/src/lib/utilities/index.ts
@@ -19,4 +19,4 @@ export * from "./useResizeObserver/index.js";
 export * from "./AnimationFrames/index.js";
 export * from "./useIntersectionObserver/index.js";
 export * from "./IsFocusWithin/index.js";
-export * from "./useFSM/index.js";
+export * from "./FiniteStateMachine/index.js";
